backend/requirements.txt

en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
Flask
Flask-Cors
Flask-JWT-Extended
huggingface-hub
loguru
numpy
presidio_analyzer
presidio_anonymizer
PyJWT
scikit-learn
scipy
sentence-transformers
intervaltree
textdistance
rapidfuzz
python-dotenv
